Flip the orientation of polygons in MVT.
The MVT docs say that:

> In a frame where the Y axis is positive up, this would make the
> winding order appear counter clockwise.

However, this doesn't mean that we can orient our polygons as
counter-clockwise, as part of the encoding process for encoding
is flipping the Y coordinate (`'y': self.extents - y`). This means
that the orientation of the polygons doesn't change as part of the
encoding process, and therefore exterior rings must be *clockwise*
when we encode them.