Sanitise attributes before outputting them.
Attributes up to this point have been arbitray Python objects, and
elsewhere we serialise them to JSON, but both Python and JSON have
a richer selection of basic types than MVT allows.

For example, transit routes are output as a `list`, which is fine
in JSON, but led to an off-by one in MVT as it was added to
`self.values` but not `layer.values` since it did not match any
clause of the `isinstance` testing.

Added test cases which check this and a few other cases. Fixed by
filtering on the types allowed for both key and value before
attempting to add either.