Fix the winding order when y is down.
The library accepts both y-up and y-down orientations for input
polygons, and attempts to re-orient them so that the winding order
is correct according to the Mapbox vector tile specification.

There was a bug in the routine for orienting the polygons, where
it did not account for whether the input polygon was already in a
y-down coordinate system and therefore should already show a
positive sign.

In addition, tests have been added based directly on the examples
in the Mapbox vector tile specification, which hopefully ensure
that the winding order behaviour is compliant.