Fix more causes of polygon invalidity.
Instead of intersecting the polygon with its reverse, just "union" the polygon on its own. This is used elsewhere and appears to be just as good as intersecting with the reverse for cleaning the polygon. It has the added benefit of being quicker, and probably less confusing to clipper.

Remove degenerate inners before calling clipper, as these seem to lead to problems when trying to convert the resulting geometry back to Shapely. It seems that there are many things which clipper considers valid, but the OGC standard disallows.

Add tests, including several culled from the logs of the dev servers.