Proper area calculation. Use PolyTree result from Clipper.
The previous area calculation always returned zero, meaning every polygon was replaced by its outer shell. This wasn't the intended behaviour! There are now many more tests to check that the expected behaviour is happening.

Clipper can return results as either a list of rings or as a PolyTree. The latter includes information about which rings are inners and which are outers, which solves some topological problems. Using this information means we can avoid certain issues.

One problem remains: When differencing valid geometries (in test `test_shapely_difference`), we expect to get a valid geometry, but we don't.