Skip inners which cause exceptions.
Differencing two polygons can sometimes result in an exception, even if both input polygons are valid. This exception would be propagated and could prevent the tile from being generated. Instead, we skip any inner for which the difference was not computable, just as we already do for any inner which causes the outer to become invalid.